move ANIResults class definition to tools module
